Separate agenda view from text rendering
