+++ caldir-cli/src/commands/mod.rs
@@ -1,3 +1,4 @@
+pub mod agenda_view;
 pub mod config;
 pub mod connect;
 pub mod discard;
